+++ pandas/core/ops/missing.py
@@ -99,7 +99,7 @@ def mask_zero_div_zero(x, y, result: np.ndarray) -> np.ndarray:
     >>> y = 0       # int 0; numpy behavior is different with float
     >>> result = x // y
     >>> result      # raw numpy result does not fill division by zero
-    array([0, 0, 0], dtype=int64)
+    array([0, 0, 0])
     >>> mask_zero_div_zero(x, y, result)
     array([ inf,  nan, -inf])
     """
